Initial commit
Copied code from EnvironmentMonitor and commented out everything related
to the LDR and LCD. I'm moving development of that project from the
Arduino to the esp8266-12e. References to the LDR were commented-out for
the time being until I get a better grasp of the I/Os for the
esp8266-12e. References to the LCD were commented out because I will
need to use a different LCD library with the esp8266-12e and having the
old references breaks the code.