Fixed accuracy of readTemperature()
Added an offset for the ADC value. The temperatures read by the
esp8266-12e were about +5C from the temperatures read by the Arduino.
After some research, I found that it is a known problem that the ADC on
NodeMCU's boards are not accurate. To compensate, I setup a voltage
divider with two 1k resistors and measured the voltage between them.
It was 1.65V, which is half of 3.3V, which is what I expected. The ADC
value was 543 when it should have been 512. To find the offset, I took
543-512 = 31 and defined the ADC_OFFSET to be 31. This brought the
temperature reading close to what I was seeing on the Arduino.
Furthermore, I changed readTemperature() to get 10 sensor readings and
find the average instead of just one reading.