Deleted printToSerial(.) and added publishData(..)
Since the networked version of this sketch does not transmit data
through serial, I removed the function that did that. In its place, I
wrote a function to publish data to the mqtt broker using pubsubclient.
I also copied a bunch of code/helper functions from the pubsubclient
esp8266 example to make it work. Yay plumbing!

TODO: change inputs to publishData function so that it accepts a
multidimensional array of topics and data instead of two arguments, one
for topic and the other for data.