Removed references to LCD. Added setup for OLED
Commented-out everything related to LCD. Copied the include statements,
constants, and setup portion from the Arducam_ssd1306_128x64_I2C.ino
example program that came with the library. Changed the position and
contents of the text so that it will print the clientID name in the
top portion of the screen where there are yellow pixels.