Added code to select channel on multiplexer
Since the esp8266 has only one analog input, I added a multiplexer IC,
the cd4051b. The temperature sensor is wired to channel 0. One problem
I ran into when setting up this chip is that I forgot to wire INH
(inhibit) to ground so that it would read a zero. Without that, the IC
does not let channel-in go to channel-out. Used a switch statement to
set the states of the digital out pins connected to the channel select
pins on the cd4051b. Added cases for channel 0 and channel 1.