Added error handling for connecting to WiFi
Set the max number of attempts to 20 because that is how many periods
will fit across the OLED and I have yet to see it take more than 15.
Created a function to handle errors called haltOnError. It accepts
a string, prints it to the serial port, displays the message to the
screen, then feeds the watchdog forever.