Added clearAndWriteToTitle(String title)
Clears the title area by writing a black rectangle to it, then writes
the title. Text size is set to 2. Changed all places in the code where
I manually udpated that section of the display to use this new function.
Causes "Temp:..." to blink in the printDataToOLED function. Hopefully
that will be fixed by functionalizing the part that writes text to the
body.