Updated code to work with mcp3008
Replaced all of the multiplexer and ADC code with calls to the mcp3008
ADC IC. Debugged hardware. Found that because the temperature sensor was
near the USB cable, it was getting bad readings, except when I connected
a multimeter to it. Moved it to the other end of the breadboard and the
mcp3008 is getting accurate readings now.