Renames main code and updates README
Renames main code to better reflect what this code actually is. The
scope of the project changed. Instead of building an environment monitor
with lots of different sensors, this is just going to be a thermostat.

Uncommented preprocessor definitions for debugging. Uncommented sections
of code to allow network connectivity again. Removed SSID, password, and
MQTT broker address from commit.

Updated README to reflect changes to project scope. Also added a Build
Instructions section and a Dependencies section.