Updates basic Wifi error handling
Previously, if the thermostat could not connect to wifi, it errored out
and hung there. Now, if it cannot connect to wifi, it will not attempt
to connect to the MQTT broker in the setup block and will not try to
send the temperature to the broker in the loop block. If the wifi is
connected and the client is connected to the data broker, it will send
the temperature to the datat broker. It periodically tries to reconnect
to the wifi network. Adds symbols to show wifi connectivity in the upper
right-hand corner of the screen based on wifi status from the last time
it checked. It does not continuously check wifi status.