Updates error handling for connecting to mqtt
Previously, if the thermostat could not connect to the mqtt broker, it
error-ed out after the max attempts had been reached. Now it will retry
to connect to the mqtt broker at a given interval. If it cannot connect,
it still reads and displays the temperature. Also added a symbol that
shows-up on the display to the left of the wifi symbol to show
whether or not the thermostat is connected to the mqtt broker. Tested
the following cases:

 - no wifi
 - wifi but no broker
 - connected to wifi and broker disappears
 - connected to wifi and wifi disappears

The thermostat was able to reconnect to wifi and the mqtt broker if
either disappeared, then reappeared while still reading the temperature
and displaying it to the OLED.