Rearranges logic for testing network connectivity
Tests mqtt broker connection before wifi. If we're connected to the mqtt
broker, then no error handling has to happen. If we're not, then it
checks the wifi connection, if we are connected to wifi but not mqtt,
then it attempts to reconnect to the mqtt broker and moves on regardless
success or failure. If the wifi is not connected, then it attempts to
reconnect to the wifi but moves on regardless of success or failure.

Adds some comments about future plans for controlling boiler based on
whether or not we're connected to the mqtt broker. If we are, then we'll
use data from it to toggle boiler. If we are not connected to the
broker, then we will use local temperature data for boiler control.