Adds local thermostat functionality
Adds the function "setBoiler" that takes a temperature as an input and
toggles the boiler based on the input temperature, the desired
temperature, and the upper and lower bounds around the desired
temperature. It will turn-on the relay pin if the temperature drops
below the desired temperature minus the lower bound. The relay pin stays
high until the temperature reaches the desired temperature plus the
upper bound, then it turns off. It stays off until the temperature drops
below the desired temperature minus the lower bound.

Tested this functionality by hard coding a desired temperature based on
the ambient temperature in the space I was working (it was pretty warm)
and holding the thermistor. Connected the relay pin to an LED to
visually see toggle. I forgot to bring my screwdriver with me so I could
not test the relay for this commit.

Also, I moved the setup and loop functions to the bottom of this program
so they are easier to find.