Fixes network connectivity bug and adds relay msg
Puts the reconnect function before testing the mqtt connection. When my
test period was only 1 second, the ESP would stay connected to the mqtt
broker between loops. However, once I started doing longer periods, it
did not. This commit runs the reconnect function before testing the mqtt
connection status.

The previous way was confusing because the indicator on the screen said
that I was connected to mqtt when in fact I was not. So I was very
confused as to why the mqtt broker was not receiving anything even
though the screen said the thermostat was connected.

This commit also adds mqtt messages to indicate the status of the relay
which implies the status of the boiler.