Moves network connectivity check inside reconnect
I was having trouble with the thermostat staying connected to the mqtt
broker. I found that I had forgotten to include the reconnect/loop
structure at the top of my loop function. This commit moves the network
connectivity check inside the reconnect function so that before trying
to connect to the mqtt broker, it first checks whether or not it is
connected to wifi. This significantly cleans-up the main loop function.

Also, this commit changes how some of the debug statements related to
network connectivity work. I move all of the serial debug print
statements into their own DEBUG blocks and moved statements that print
to the OLED into DEBUG_ON_OLED blocks.