Starts refactor. Adds Thermometer.cpp. Guts .ino
The original file, NetworkedThermostat.ino has become unreadable so a
refactor is in order. This commit starts that process. The goal of the
refactor is to make the code more modular. Since the room nodes will
share a large code base with the thermostat, it is prudent to refactor
the code in such a way as to make that easy. At the end of this
refactor, there will be two main objects: Thermometer and Thermostat.

This commits starts the Thermometer object. Thermometer.ccp is a class
file. This commit adds the ability to create a Thermometer object,
connect the esp8266 to the si7021 chip via I2C, and read the ambient
temperature. NetworkedThermostat.ino is updated to create a
Thermometer object, read the ambient temperature, and print it to
Serial.