Adds header files for custom classes and cleans-up code
NOTE: I did not run the code for this commit on an esp8266 module. I
only verified that it compiles. I haven't touched this project in
almost a year and a half and no longer am running the mqtt/database
server to which this code connected.

I wrote these classes before I really understood how header files and
classes worked. The code ran, so I was happy at the time. Since then,
I've grown a lot as a programmer so I decided to update this repo to
reflect that.

Squashed commit of the following:

commit c3bd091412a321929d1105203fa8e0519b50a3cd
Author: Peter Mowen <pmmowen@gmail.com>
Date:   Sun Dec 5 12:37:12 2021 -0500

    Removes commented-out code from NetworkedThermometer.ino

commit 90aab57aafd1d06ca1a966c894058e3b5afe4de9
Author: Peter Mowen <pmmowen@gmail.com>
Date:   Sun Dec 5 12:36:23 2021 -0500

    Removes commented-out code from HomeNode.cpp

commit 84089ba8abf7ca1d1c80421ea656abd3c467e70a
Author: Peter Mowen <pmmowen@gmail.com>
Date:   Sun Dec 5 12:21:14 2021 -0500

    Adds guard clauses to header files. Cleans-up NetworkedThermometer.ino

commit cdff0749c53320f2b5440506f6258e0e26906a6e
Author: Peter Mowen <pmmowen@gmail.com>
Date:   Sun Dec 5 12:16:17 2021 -0500

    Adds Thermometer.h and updates files to use it

commit 07b5eaf102a6687f8e061d92b92b2e80d3aadf71
Author: Peter Mowen <pmmowen@gmail.com>
Date:   Sun Dec 5 11:55:50 2021 -0500

    Adds a header file for Si7021 and updates Thermometer.cpp to use it

commit 92425fa50061ea1e228c137ac87718e3f46b7b6c
Author: Peter Mowen <pmmowen@gmail.com>
Date:   Sun Dec 5 11:48:24 2021 -0500

    Adds HomeNode.h and fixes typo in NetworkedThermometer.ino file name

    When I originally wrote this, I did not understand header files. This
    was my first attempt at understanding C++ classes. I was kind of a dope.